+++ kubernetes-core.el
@@ -37,13 +37,30 @@
         ;; Suppress redrawing if the overview is not selected. This prevents
         ;; point from jumping around when a magit popup is open.
         (when (member (selected-window) (get-buffer-window-list buf))
-          (kubernetes-utils--save-window-state
+          (kubernetes--save-window-state
            (let ((inhibit-read-only t))
              (erase-buffer)
              (kubernetes-ast-eval (kubernetes--overview-render (kubernetes-state)))))
 
           ;; Force the section at point to highlight.
           (magit-section-update-highlight))))))
+
+
+(defmacro kubernetes--save-window-state (&rest body)
+  "Restore window state after executing BODY.
+
+This is useful if the buffer is erased and repopulated in BODY,
+in which case `save-excursion' is insufficient to restore the
+window state."
+  `(let ((pos (point))
+         (col (current-column))
+         (window-start-line (window-start))
+         (inhibit-redisplay t))
+     (save-excursion
+       ,@body)
+     (goto-char pos)
+     (move-to-column col)
+     (set-window-start (selected-window) window-start-line)))
 
 (defun kubernetes--message (format &rest args)
   "Call `message' with FORMAT and ARGS.